fix the so far broken FIM_USE_CXX11 definition
Otherwise even if set, it would lead to FIM_IMG_NAKED_PTRS to expand to (!1|| 1)=1, instead of (!(1||1))=0.
That in turn would lead naked pointers to be used in non-pre-c++11 code, with all the consequent bugs.
Set a deprecation message next to FIM_DEPRECATED-implied usage, just in the Cache class declaration.